Developed Brain Wrapper
Created the Brain class, which is basically a wrapper for everything
else. The reason for implementing this was so that I could have the
freedom to change names and organize without breaking everything. Also,
this means the Sensor and interface methods will never need to be
looked at as long as it is know that a few of the variables are updated
automatically. This is designated with comments. Definitely not “clean”
but it gets the job done and its better then the alternative for me.
Having full control over every thing is valuable to me.