Fixed a couple bugs
Specifically, I fixed the spiral path to have the right number of
nodes. Additionally, I hacked together a way to check if the black hole
is being routed to. This will need to change, its not sustainable but
will allow me to figure out some of the more pressing matters first.
The next issue to fix si that the reference to ScoutControl in Nav is
not propagating to Brain