Broke up ScoutControl and ScoutStatus into classes
It turns out the problem was that structs are value types while classes
are reference types. So my attempts to pass the ScoutControl reference
along were impossible. So I broke them up into classes so I can use
them as references. The potential downside to this is that a new class
has to be built each frame for each of these instead of just assigning
them a new value. It doesn’t seem to be affecting the speed at all
though. The next issue is that the scout is moving in a random manor. I
expect the algorithm copied from the professor got changed somewhere.