Figured out Avoidance problem, partial implementation
The problem I was having was that the PID was being calculated every
time MoveToTarget  or MoveAwayFromTarget was called. This produced
crazy results since PID’s rely on previous PID’s to generate error. So
I moved the PID calculations to their own method. Now, the pid is
calculated after the target thrust is returned from the MoveToTarget
method. This is done in the Brain class, as it fits in more there.
Additionally, I’ve started to use ScoutThrustControls for storing
prescribed thrusts instead of pose in order for the representation to
be accurate and more readable. Additionally, the navigation class has 2
less references. Decoupling was a major improvement in this commit. The
next thing to do will be to implement object avoidance and weight
everything in the think method.