Refined object avoidance
Object avoidance is now mainly concerned with objects directly in its
path and objects within critical range. side thrusters are used for
everything non-critical. A range of 2/3pi in front of the  scout is
used to categorize objects as “in the way” or not. This implementation
fares well in the main game even without wall avoidance.