Orbit is functional!
I’ve been thinking it was a problem with track ship but the actual
issue was that I wasn’t calculating the PID after my orbit calculation.
This is why hacking together things is dangerous! My next move will be
cleaning up the method, and playing around with the javelin idea again
now that I know it actually works. Also I’ll need to define a better
condition for when it can leave orbit .