Existential crisis commit
Cant get shit to work. I added a safety net for the scout. If it goes
out of bounds, it routes to the center. I commented out the averages
and the black hole crit thing for now as it doesn’t seem to do much.
I’m thinking maybe since I fixed that Y coord, moving the forward
thrust out of the cone might be more useful